Updated Ubuntu 12.04 instructions
This was tested on the same machine as before. Changes are:

- xmlto is required (it was already installed on my machine so I missed this)

- libinput is enabled by default

- fixed date

Several x11 proto packages that seem to be required on my Mint 14.04 machine
are not necessary here: randrproto, compositeproto, scrnsaverproto,
xineramaproto.

I believe the reason is that the libx11-dev package necessary for mesa
installed sufficiently-recent versions.

(based on patch by: Jon A. Cruz <jon@joncruz.org>)
Reviewed-by: Bryce Harrington <b.harrington@samsung.com>