doc: Update ubuntu12.04 instructions to build expat from source
Uses the tar file from the sourceforge repository. This avoids use of cvs
which would otherwise have to be installed.

The main purpose of this is to get a working pkg-config so the new
autoconf for wayland works, but since this updates from expat 1.6.0
to 2.1.0 it seems like a good idea.