Drop the ubuntu 12.04 build directions.
It is unlikely anyone still needs directions on how to install on this
old distro -- Ubuntu 12.04 is scheduled to hit end-of-life this April.

Further, no developers (to my knowledge) still test on 12.04, so the
directions have likely bitrotted anyway.  (Bill had a machine running
12.04 and was keeping the page updated, but last March indicated in
4fa80f28 he is no longer testing on it.  I myself moved off 12.04 some
time ago as well.)

For the most part, the directions are requiring a nearly full build of
the stack from source, thus is rather duplicative of the generic Wayland
build directions (which are more actively maintained).  Only a handful
of lower level X packages and some compiler tools are used from the
system.

Signed-off-by: Bryce Harrington <bryce@osg.samsung.com>
Acked-by: Daniel Stone <daniels@collabora.com>